feat(server): enable new dashboard design by default
