+++ packages/server/src/ui/routes/project-dashboard/graphs/category-score/category-score-graph.jsx
