+++ packages/server/src/ui/routes/project-dashboard/graphs/category-score/score-delta-bar-graph.jsx
