+++ packages/server/src/ui/routes/project-dashboard/graphs/category-score/score-distribution-graph.jsx
