+++ packages/server/src/ui/routes/project-dashboard/graphs/category-score/score-line-graph.jsx
