+++ packages/server/src/ui/routes/project-dashboard/graphs/metric-line-graph.stories.jsx
