+++ packages/server/src/ui/routes/project-dashboard/project-category-summaries.jsx
