--- packages/server/src/ui/routes/project-dashboard/project-graphs.css
@@ -1,73 +1,0 @@
-/**
- * @license Copyright 2019 Google Inc. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
- */
-
-.dashboard-graphs {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-}
-
-.dashboard-graphs__legend {
-  width: 100%;
-  padding: var(--base-spacing);
-
-  font-weight: var(--medium-font-weight);
-
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-.dashboard-graphs__legend-items {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
-
-  margin-right: var(--base-spacing);
-}
-
-.legend-item {
-  margin: 0 var(--base-spacing);
-
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-.legend-item__color-chip {
-  height: 2px;
-  width: calc(2 * var(--base-spacing));
-  margin-right: calc(var(--base-spacing) / 2);
-}
-
-.dashboard-graphs__branch-select,
-.dashboard-graphs__branch-select select {
-  max-width: 200px;
-}
-
-.dashboard-graph {
-  width: calc(50% - 2 * var(--base-spacing));
-  /* Note this value must be set in JavaScript as well for the Plot height. */
-  min-height: var(--dashboard-graph-min-height);
-  overflow: hidden;
-}
-
-.dashboard-graph__title {
-  margin-top: var(--base-spacing);
-  margin-left: var(--base-spacing);
-}
-
-.dashboard-graph__plot {
-  /*
-   * Plots have the x-axis on the right which creates weird empty space on the left.
-   * This adjusts for it by removing margin and shifting the element outside the paper.
-   */
-  margin-left: calc(var(--paper-padding) * -2);
-}
