+++ packages/server/test/e2e/steps/navigate-to-project.js
@@ -9,7 +9,7 @@ const {waitForNetworkIdle0, waitForAllImages} = require('../../test-utils.js');
 
 /* eslint-env jest, browser */
 
-/** @param {LHCI.E2EState} state @param {string} projectName @param {{newDashboard: boolean}} [opts] */
+/** @param {LHCI.E2EState} state @param {string} projectName @param {{}} [opts] */
 module.exports = (state, projectName, opts) => {
   describe('navigate to project', () => {
     it('should navigate to the project list', async () => {
@@ -29,31 +29,8 @@ module.exports = (state, projectName, opts) => {
     });
 
     it('should wait for the dashboard to load', async () => {
-      await state.page.waitFor('.dashboard');
-      await state.page.waitFor('.dashboard-graph');
+      await state.page.waitFor('.category-score-graph');
     });
-
-    if (opts && opts.newDashboard) {
-      it('should navigate to new dashboard', async () => {
-        const clickPromise = state.page.evaluate(() => {
-          const links = [...document.querySelectorAll('a')];
-          const matchingLink = links.find(link =>
-            (link.textContent || '').includes('redesigned dashboard')
-          );
-          if (matchingLink) return matchingLink.click();
-          throw new Error(
-            `${links.map(link => link.textContent).join(' ')} did not include ${name}`
-          );
-        });
-
-        await Promise.all([clickPromise, waitForNetworkIdle0(state.page)]);
-        await waitForAllImages(state.page);
-      });
-
-      it('should wait for the dashboard to load', async () => {
-        await state.page.waitFor('.category-score-graph');
-      });
-    }
 
     if (state.debug) {
       it(
